requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.3
APScheduler==3.7.0
asgiref==3.3.4
async-timeout==3.0.1
attrs==19.3.0
audio-metadata==0.11.1
Automat==0.8.0
beautifulsoup4==4.9.3
bidict==0.21.2
bitstruct==8.11.1
blinker==1.4
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.7
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.3
django-admin-list-filter-dropdown==1.0.3
django-filter==2.4.0
djangorestframework==3.12.4
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
Faker==8.1.0
filelock==3.0.12
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jwt==1.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.4
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==5.1.0
neo4j-driver==4.1.1
neobolt==1.7.17
neomodel==4.0.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pendulum==2.1.2
pexpect==4.6.0
Pillow==7.0.0
pluggy==0.13.1
pprintpp==0.4.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
pytest==6.2.3
python-apt==2.0.0
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.17.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
Shapely==1.7.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
tbm-utils==2.6.0
text-unidecode==1.3
texttable==1.6.2
toml==0.10.2
Twisted==18.9.0
twitchio==1.2.3
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
websocket-client==0.58.0
websockets==8.1
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.4.26
zipp==1.0.0
zope.interface==4.7.1
